refactor(ui): split TestResultModal into summary/details/logs tabs
